--- Dayana/Server/Api/ResultFilters/Identity/Auth/TokenResultFilter.cs
@@ -1,21 +1,0 @@
-﻿using Dayana.Shared.Persistence.Models.Identity.Base;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Identity.Auth;
-
-public class TokenResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is TokenResult value)
-            result.Value = new
-            {
-                value.AccessToken
-            };
-
-        await next();
-    }
-}
