--- Dayana/Server/Api/ResultFilters/Identity/Permissions/GetPermissionsResultFilter.cs
@@ -1,34 +1,0 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
-using Dayana.Shared.Infrastructure.Pagination;
-using Dayana.Shared.Persistence.Models.Identity.Base;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Identity.Permissions;
-
-public class GetPermissionsResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is PaginatedList<PermissionModel> value)
-            result.Value = new
-            {
-                value.Page,
-                value.PageSize,
-                value.TotalCount,
-                Data = value.Data.Select(x => new
-                {
-                    Eid = x.Id.EncodeInt(),
-                    x.Title,
-                    x.Name,
-                    x.Value,
-                    x.CreatedAt,
-                    x.UpdatedAt
-                })
-            };
-
-        await next();
-    }
-}
