--- Dayana/Server/Api/ResultFilters/Identity/Roles/DeleteRoleResultFilter.cs
@@ -1,22 +1,0 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
-using Dayana.Shared.Domains.Identity.Roles;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Identity.Roles;
-
-public class DeleteRoleResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is Role value)
-            result.Value = new
-            {
-                Eid = value.Id.EncodeInt()
-            };
-
-        await next();
-    }
-}
