--- Dayana/Server/Api/ResultFilters/Identity/Users/CreateUserPermissionResultFilter.cs
@@ -1,23 +1,0 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
-using Dayana.Shared.Domains.Identity.Claims;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Identity.Users;
-
-public class CreateUserPermissionResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is Claim value)
-            result.Value = new
-            {
-                Eid = value.Id.EncodeInt(),
-                value.Value
-            };
-
-        await next();
-    }
-}
