--- Dayana/Server/Api/ResultFilters/Identity/Users/UpdateUserRolesResultFilter.cs
@@ -1,23 +1,0 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
-using Dayana.Shared.Domains.Identity.Users;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Identity.Users;
-
-public class UpdateUserRolesResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is UserRole value)
-            result.Value = new
-            {
-                RoleId = value.RoleId.EncodeInt(),
-                UserId = value.UserId.EncodeInt()
-            };
-
-        await next();
-    }
-}
